update tsconfig and package.json (#3)
## Description

This PR updates the `tsconfig.json` and `package.json` files. With regard to the `tsconfig.json` it adds multiple compiler options that generate declaration files (`.d.ts`), introduce stricter compiler rules, define the module resolution and other smaller things. The changes in the `package.json` are quite trivial. There, it only updates the repo links and adds a pointer to the declaration files.

## Commits 

* update tsconfig and package.json